requirements.txt
Flask
requests
pymongo
mysql-connector-python
psycopg2
